Restructure files, add SASS, implement new styles and colors to page layout
