--- src/Movie.css
@@ -1,24 +1,0 @@
-.movie-card {
-    height: 100%;
-    width: 100%;
-}
-
-.movie-card-image {
-    width: 70%;
-    height: 80%;
-}
-
-.movie-card-details {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-}
-
-.title-year {
-    display: flex;
-    width: 100%;
-    justify-content: space-evenly;
-    margin: 0;
-}
-
