--- src/MoviesArea.css
@@ -1,10 +1,0 @@
-.movie-container {
-    margin-top: 5%;
-    height: 80%;
-    width: 100%;
-    display: grid;
-    overflow-y: auto;
-    grid-template-columns: repeat(3, 1fr);
-    margin: 0;
-}
-
