+++ src/components/MoviesArea.js
@@ -1,5 +1,5 @@
 import Movie from './Movie';
-import './MoviesArea.css';
+import '../styles/MoviesArea.scss';
 function MoviesArea({ movies }){
     const allMovies = movies.map(movie => {
         return (
