+++ src/components/Navbar.js
@@ -1,4 +1,4 @@
-import './Navbar.css';
+import '../styles/Navbar.scss';
 
 export default function Navbar() {
     return (
