+++ src/index.js
@@ -1,8 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './index.css';
-import App from './App';
+import './styles/index.scss';
+import App from './components/App';
 import reportWebVitals from './reportWebVitals';
+import './assets/star-icon.svg';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
