+++ src/styles/MoviesArea.scss
@@ -1,0 +1,6 @@
+@import 'mixins';
+
+.movie-container {
+    @include movie-area();
+}
+
