+++ src/styles/Navbar.scss
@@ -1,10 +1,10 @@
+@import 'mixins';
+
 .nav {
     height: 20%;
     width: 100%;
     margin: 0;
     margin-bottom: 2%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
     text-align: center;
+    @include flex-centered-row();
 }
